# porting single-compress optimization to arm