@WojciechMula recently pushed a nice optimization to the Westmere and Haswell kernels. We can port it to ARM.

Here are the results on my M2 macBook.

Command:
```
cmake --build build --target benchmark_base64  && ./build/benchmarks/base64/benchmark_base64 -d ../base64data/dns/swedenzonebase.txt
```

Before:
```
simdutf::arm64                           :   5.52 GB/s  2.51 % 
simdutf::arm64 (accept garbage)          :   5.70 GB/s  1.78 % 
```

After:
```
simdutf::arm64                           :   6.20 GB/s  3.11 % 
simdutf::arm64 (accept garbage)          :   6.86 GB/s  2.43 % 
```

So that's a performance boost of about 10%... which is well worth it.

The downside of this approach is more branching and slightly more complexity, so it is possible that the gains do not always translate into benefits. However, we do expect the branches to be predictible because inputs are going to be regular in some ways (in practice).